mysite/calculator/api/urls.py
from django.urls import include, path
from rest_framework import routers

from .views import PersonViewSet

router = routers.SimpleRouter()
router.register('persons', PersonViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
]

urlpatterns += router.urls
